add more tests to e2e lotto

+++ SadnaSrc/SadnaSrc/Main/MarketDB.cs
@@ -314,7 +314,7 @@ namespace SadnaSrc.Main
             return @"CREATE TABLE IF NOT EXISTS [Products] (
                                     [SystemID]     TEXT,
                                     [Name]         TEXT,
-                                    [BasePrice]    INTEGER,
+                                    [BasePrice]    REAL,
                                     [Description]  TEXT,
                                     PRIMARY KEY([SystemID])
                                     )";
@@ -349,8 +349,8 @@ namespace SadnaSrc.Main
             return @"CREATE TABLE IF NOT EXISTS [LotteryTicket] (
                                     [myID]              TEXT,
                                     [LotteryID]         TEXT,
-                                    [IntervalStart]     INTEGER,
-                                    [IntervalEnd]       INTEGER,
+                                    [IntervalStart]     REAL,
+                                    [IntervalEnd]       REAL,
                                     [Cost]              REAL,
                                     [Status]            TEXT,
                                     [UserID]            INTEGER,
@@ -364,8 +364,8 @@ namespace SadnaSrc.Main
             return @"CREATE TABLE IF NOT EXISTS [LotteryTable] (
                                     [SystemID]              TEXT,
                                     [ProductSystemID]       TEXT,
-                                    [ProductNormalPrice]    INTEGER,
-                                    [TotalMoneyPayed]       INTEGER,
+                                    [ProductNormalPrice]    REAL,
+                                    [TotalMoneyPayed]       REAL,
                                     [storeName]             TEXT,
                                     [StartDate]             TEXT,
                                     [EndDate]               TEXT,
